In studentModel, there was a bug in the mongoose middleware for join i.e. currentStatus was typed as active, I changed it to currentStatus. In config.env, the variable for cookie expire time was added i.e. COOKIE_EXPIRE_IN.
In authController, create token and sending the response has been added in the function and instead of writting the code everywhere,
simply that method is called. To hide the password in the output i.e. when signing up or logining in the student.password value is definded as undefined so it doesn't appear in the response output